[#97] fix : lecture에 처리했던 부분을 EnrollmentService로 이동
